requirements.txt
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.5
django-suit==0.2.25
django-toolbelt==0.0.1
djangorestframework==3.5.4
gunicorn==19.6.0
MySQL-python==1.2.5
psycopg2==2.7.1
wheel==0.24.0
whitenoise==3.3.0
factory-boy==2.7.0
fake-factory==0.5.10